all: a number of updates to accomodate improved design understanding

+++ afsk/radio.go
@@ -3,7 +3,7 @@ package afsk
 // Radio defines the interface for AFSK radio transmitters.
 type Radio interface {
 	// Transmit sends a signal at the specified frequency.
-	Transmit(freq uint32) error
+	Transmit(freq uint64) error
 	// Standby puts the radio into standby mode.
 	Standby() error
 	// Close releases resources associated with the radio.
